Add Settings and Application Search
These two actions require us to manipulate and read plist values. To
do this, we're using Rubycocoa on the client.  Now, there's a stupid
Live Management issue here - if you copy the agent to a master running
Linux, and that master DOESN'T have Rubycocoa, then it refuses to load
the agent AND the Controlling Puppet and Advanced Tasks section of the
Puppet Enterprise Console become completely blank.  Yeah. That sucks.
Instead, I rescue the LoadError, comment the use case, and do a check
for the file so it doesn't get attempted on Linux Boxes.